Clean up temp directory on exit, even if current directory
It's quite tempting to do something like this:

    use File::Temp qw(tempdir);
    my $dir = tempdir(CLEANUP => 1);
    chdir $dir or die;
    # now do some work in the current directory

However, the directory isn't automatically cleaned up because you cannot
rmtree() the current directory. Recent releases of File::Temp and
File::Path give a warning about this, but wouldn't it be better for it
to just do the right thing?

This patch makes the directory be cleaned up on exit by doing chdir('/')
if necessary. That happens only for automatic cleanup at END - if you
manually call cleanup() in the middle of your program it is still your
responsibility to sort out the current directory.

(This is from <http://rt.cpan.org/Public/Bug/Display.html?id=45246>)