Add test descriptions and use rel2abs on tempdir return
The test was passing erroneously because there was a chdir at the
end of the test that meant the directories were not being found
properly (and since they were not found the test passed despite
the directories still being present). Now the temp directories
are stored with a full path and with this commit the tests will
fail properly.