Safer way to manage real paths and untainting
We now give back relative pathnames for directories.  This avoids an
attacker running from a directory with malicious character sequences
and giving an absolute path to the user that they might inadvertantly
use.

However, we save the untainted absolute path for later unlinking because
we know how we're using it for deletion.